[Decomp][Player] Cohesion split: extract rested-XP system to PlayerRest.cpp
